almost done

+++ ft_strncat.c
@@ -6,7 +6,7 @@
 /*   By: yuxu <marvin@42.fr>                        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/25 16:21:47 by yuxu              #+#    #+#             */
-/*   Updated: 2017/11/25 18:37:23 by yuxu             ###   ########.fr       */
+/*   Updated: 2017/12/02 20:41:36 by yuxu             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ char	*ft_strncat(char *dest, const char *src, size_t n)
 	m = 0;
 	while (dest[i])
 		i++;
-	while (src[m] && m < n)
+	while (src[m] && m < (int)n)
 	{
 		dest[i] = src[m];
 		i++;
